implemented BackgroundGradientContainer in homePage
